admin/requirements.txt
aiofiles==23.2.1
aioredis==2.0.1
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cycler==0.11.0
exceptiongroup==1.2.0
fastapi==0.103.2
fastapi-admin==1.0.4
fonttools==4.38.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.6
importlib-metadata==6.7.0
iso8601==1.1.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.9.0.80
packaging==23.2
pandas==1.3.5
pendulum==2.1.2
Pillow==9.5.0
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.6
python-telegram-bot==20.3
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
pyzbar==0.1.9
qrdet==2.4
qreader==3.12
quadrilateral-fitter==1.8
requests==2.31.0
scipy==1.7.3
seaborn==0.12.2
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
torch==1.13.1
torchvision==0.14.1
tortoise-orm==0.19.3
tqdm==4.66.1
typing_extensions==4.7.1
ultralytics==8.0.145
urllib3==2.0.7
uvicorn==0.22.0
uvloop==0.18.0
watchfiles==0.20.0
websockets==11.0.3
zipp==3.15.0